Rename fuzzy_typeahead.ts file to use_fuzzy_typeahead.ts
This is more clear, since it follows the pattern of other files in this
directory and matches the name of the primary named export in this file.
