+++ app/javascript/lib/composables/use_fuzzy_typeahead.ts
